[front] Refactor `GetMCPActions` to use `AgentStepContentResource` and optimize cursor handling
