+++ front/pages/api/w/[wId]/labs/mcp_actions/[agentId]/index.ts
@@ -4,25 +4,21 @@ import * as reporter from "io-ts-reporters";
 import { NumberFromString, withFallback } from "io-ts-types";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import type { ToolExecutionStatus } from "@app/lib/actions/statuses";
 import { getAgentConfiguration } from "@app/lib/api/assistant/configuration/agent";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
 import { AgentMCPActionResource } from "@app/lib/resources/agent_mcp_action_resource";
+import { AgentStepContentResource } from "@app/lib/resources/agent_step_content_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types";
+import type { AgentMCPActionType } from "@app/types/actions";
 
 export type GetMCPActionsResult = {
-  actions: {
-    sId: string;
-    createdAt: string;
-    functionCallName: string | null;
-    params: Record<string, unknown>;
-    status: ToolExecutionStatus;
+  actions: AgentMCPActionType[] & {
     conversationId: string;
     messageId: string;
-  }[];
+  };
   nextCursor: string | null;
   totalCount: number;
 };
@@ -67,6 +63,20 @@ async function handler(
 
       const { agentId, limit, cursor } = queryValidation.right;
 
+      let cursorDate: Date | undefined;
+      if (cursor) {
+        cursorDate = new Date(cursor);
+        if (isNaN(cursorDate.getTime())) {
+          return apiError(req, res, {
+            status_code: 400,
+            api_error: {
+              type: "invalid_request_error",
+              message: "Invalid cursor format.",
+            },
+          });
+        }
+      }
+
       const owner = auth.getNonNullableWorkspace();
       const flags = await getFeatureFlags(owner);
       if (!flags.includes("labs_mcp_actions_dashboard")) {
@@ -106,37 +116,23 @@ async function handler(
         });
       }
 
-      const result = await AgentMCPActionResource.listByAgent(auth, {
-        agentConfiguration,
-        limit,
-        cursor,
+      const { stepContents, totalCount, nextCursor } =
+        await AgentStepContentResource.listFunctionCallsForAgent(auth, {
+          agentConfiguration,
+          limit,
+          cursor: cursorDate,
+        });
+
+      const actions = await AgentMCPActionResource.fetchByStepContents(auth, {
+        stepContents,
       });
 
-      if (result.isErr()) {
-        const error = result.error;
-        if (error.message === "Invalid cursor format") {
-          return apiError(req, res, {
-            status_code: 400,
-            api_error: {
-              type: "invalid_request_error",
-              message: "Invalid cursor format.",
-            },
-          });
-        }
-
-        return apiError(req, res, {
-          status_code: 500,
-          api_error: {
-            type: "internal_server_error",
-            message: "Failed to fetch MCP actions from database.",
-          },
-        });
-      }
-
-      const { actions, nextCursor, totalCount } = result.value;
-
       return res.status(200).json({
-        actions,
+        actions: actions.map((a) => ({
+          ...a.toJSON(),
+          conversationId: a.stepContent.agentMessage.message.conversation.sId,
+          messageId: a.stepContent.agentMessage.message.sId,
+        })),
         nextCursor,
         totalCount,
       });
