+++ front/types/actions.ts
@@ -6,16 +6,20 @@ import type { ActionGeneratedFileType } from "@app/lib/actions/types";
 import type { ModelId } from "@app/types/shared/model_id";
 
 export type AgentMCPActionType = {
+  id: ModelId;
+  sId: string;
+  createdAt: number;
+
   agentMessageId: ModelId;
-  citationsAllocated: number;
+  internalMCPServerName: InternalMCPServerNameType | null;
+  mcpServerId: string | null;
   // TODO(MCPActionDetails): prevent exposing the function call name
   //  currently used in the extension to guess the tool name but quite brittle.
   functionCallName: string;
   functionCallId: string;
-  id: ModelId;
-  internalMCPServerName: InternalMCPServerNameType | null;
-  mcpServerId: string | null;
+
   params: Record<string, unknown>;
+  citationsAllocated: number;
   status: ToolExecutionStatus;
   step: number;
 };
